errors: Support also behaviour functions in custom error types
If the underlying type of an error implements a function like

    interface{ BehaviourName() bool }

Then the matcher can assert and check if the type has this specific error
behaviour. This allows other packages to implement an error behaviour without
even importing this package. The matching of the interface implementation has
been shifted to function MatchInterface because of the performance penalties
when not matching an error behaviour. Matching a Kind error is >40x faster than
matching a specific interface implementation error.